requirements.txt
flask==3.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
sumy==0.11.0
numpy==1.26.4
nltk==3.8.1
scikit-learn==1.4.1.post1
spacy==3.7.4
networkx==3.2.1
cohere==4.51
